LDE-174: moved write tests
